move logic to create calculated data documents to a separate file envision
